extract the scroll function into utils
